Control: change is_readable/is_writable to is_writeonly/is_readonly.
These two methods were not well-thought. While a control that has the
read-only flag set is certainly not writeable, a control with the
write-only flag set could still be unwriteable because of being inactive.

The original intention for these two methods was to check for the flags,
so adjust their behaviour and naming accordingly.