Control: reimplement method is_writeable, with correct behaviour.
Before 000696e, a control was assumed (and reported by this method) to be
writeable when the read-only was not set. However, such a control would
still be unwriteable when being flagged as inactive, which was not taken
into account before.

In other words, is_writeable now checks that a control can *actually* be
written to.