Add support for checking all currently defined control flags.
is_<flag> is working for many of the flags, but is weird for stuff like
V4L2_CTRL_FLAG_HAS_PAYLOAD or V4L2_CTRL_FLAG_UPDATE. Switching to
is_flagged_<flag> to check if a flag is set, e.g. is_flagged_has_payload,
is still not perfect, but better.