Initial implementation of class MenuControl.
This class inherits from UserDict and thus can be used like a dict to access
the items defined for the menu. For the sake of this dict-like behaviour the
MenuItem class is not used in MenuControl; the names of the menu items will
be transformed straight to either string or integer, depending on the menu
type.

MenuItem is kept for now, but renamed to LegacyMenuItem to signal it should
not be used except for legacy code.